requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==4.2.17
django-apscheduler==0.7.0
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-db-geventpool==4.0.7
django-js-asset==2.2.0
django-modeltranslation==0.19.12
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
frozenlist==1.4.1
future==1.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
msgpack==1.1.0
multidict==6.0.5
mysql-connector-python==9.0.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.0.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
yarl==1.11.0
